Add support for the case that RDoc is installed as a default gem
